Rename chat.tsx to ChatBox.tsx
