+++ app/[channelId]/ChatBox.tsx
